pyproject.toml
[tool.poetry]
name = "02-advanced-chatbot"
version = "0.1.0"
description = ""
authors = ["julio colomer <info@aiaccelera.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
langchain = "^0.2.11"
langchain-openai = "^0.1.20"
python-dotenv = "^1.0.1"
jupyterlab = "^4.2.4"
langchain-community = "^0.2.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
